modiffier le frontend

+++ backend/Controllers/userControllers.js
@@ -1,7 +1,7 @@
 const User = require('../Models/User.js');
 const jwt = require('jsonwebtoken');
 const bcrypt = require('bcrypt');
-const fs = require('fs')
+const fs = require('fs');
 
 
 // fonction pour s'inscrire
@@ -16,19 +16,42 @@ module.exports.signup = (req, res, next) => {
         lastName: req.body.lastName,
         password: hash,
       })
-
       User.create(user, (err, data) => {
-
-
         if (err) {
           return res.status(400).json({ message: 'Impossible de créer l\'utilisateur' });
         }
         res.send(data);
       })
-
     })
     .catch(error => res.status(500).json({ error }));
 };
+/*
+module.exports.signup = (req, res, next) => {
+  bcrypt.hash(req.body.password, 10)// hacher le mot de passe recupere de la requette
+    .then(hash => {
+      const user = new User({ // cree  un nouveau utilisateur  avec le mot de passe hacher           
+        pseudo: req.body.pseudo,
+        email: req.body.email,
+        firstName: req.body.firstName,
+        lastName: req.body.lastName,
+        password: hash,
+      })
+      User.findOneByEmail(req.body.email, (err, user) => {
+        if (err) {
+          
+          User.create(user, (err, data) => {
+            if (err) {
+              return res.status(400).json({ message: 'Impossible de créer l\'utilisateur' });
+            }
+            res.send(data);
+          })
+        }return res.status(200).json({ message: 'user exixt deja' });
+        
+    })
+    .catch(error => res.status(500).json({ error }));
+}*/
+  
+  
 
 
 // function pour se connecter
@@ -50,12 +73,12 @@ module.exports.login = (req, res, next) => {
           } else {
             res.status(200).json({
               userId: user.id,
-              isAdmin : user.isAdmin,
+              isAdmin: user.isAdmin,
 
               token: jwt.sign({
                 userId: user.id,
-                isAdmin : user.isAdmin,
-                
+                isAdmin: user.isAdmin,
+
               },
                 process.env.TOKEN,
                 {
@@ -104,11 +127,11 @@ exports.getOneUser = (req, res, next) => {
 
 
 
-// modifier user
+// modifier l'image  user
 exports.updateOneUser = (req, res, next) => {
-  const user =  {
-    'id': req.params.id,    
-    'imageUrl': req.file ? `${req.protocol}://${req.get('host')}/images/${req.file.filename}`: null,
+  const user = {
+    'id': req.params.id,
+    'imageUrl': req.file ? `${req.protocol}://${req.get('host')}/images/${req.file.filename}` : null,
   }
   User.modifyImage(user, (err, result) => {
     if (err) {
@@ -120,11 +143,12 @@ exports.updateOneUser = (req, res, next) => {
   })
 };
 
+// modifier mail et pseudo   user
 exports.updateUserText = (req, res, next) => {
-  const user =  {
-    'id': req.params.id,    
+  const user = {
+    'id': req.params.id,
     'pseudo': req.body.pseudo,
-    'email':req.body.email,   
+    'email': req.body.email,
   }
   User.modifyUser(user, (err, result) => {
     if (err) {
@@ -141,12 +165,16 @@ exports.updateUserText = (req, res, next) => {
 // fonction supprimer le user
 
 exports.deleteOneUser = (req, res, next) => {
+  if (req.isAdmin) {
+    User.deleteUser(req.params.id, (err, result) => {
 
-  User.deleteUser(req.params.id, (err, result) => {
+      if (err) {
+        return res.status(400).json({ message: 'Impossible de supprimer l\'utilisateur' });
+      }
+      res.status(200).json({ message: 'User supprimer' });
+    })
+  } else {
+    res.status(401).json({ message: 'vous avez pas le droit de supprimer ce compte' });
 
-    if (err) {
-      return res.status(400).json({ message: 'Impossible de supprimer l\'utilisateur' });
-    }
-    res.status(200).json({ message: 'User supprimer' });
-  })
+  }
 };
